feat: bot now automatically send message data to backend whenever there is a new one
